Created linegraphs
Five linegraphs and a legend now exist in a flexbox below the violinplot. The linegraphs each represent the same data shown in each of the five violins, but show actual dollar values instead of a percent of the total.

Needs to be converted into dollars per FTE.

Labels on the x-axis need to be fixed.

Colors need to be changed.